Refactor CSS structure and styles; rename and consolidate CSS files, update imports, and remove unused components
