+++ src/App.css
@@ -1,4 +1,5 @@
 body {
+  margin: 0;
   background-color: black;
   color: white;
 }
