+++ src/css/Create_Page.css
@@ -1,4 +1,4 @@
-h2 {
+.DivCreateMain h2 {
     color: white;
     font-size: 2em;
     text-align: center;
