+++ src/css/Main_Page.css
@@ -1,4 +1,4 @@
-body {
+.Main body {
     margin: 0;
     padding: 0;
     background-color: #222222;
