+++ src/pages/Create_Page.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
-import '../Create_Page.css';
+import '../css/Create_Page.css';
 
 const Create_Page = () => {
     const { id_categorie, id_page } = useParams();
