--- src/pages/home.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-function home() {
-    return (
-        <div>
-            <h1>Hello world !</h1>
-        </div>
-    );
-}
-
-
-export default home;
-
